Renamed 'kevin' To Uppercase It
